+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 - TODO idmatcher
-- TODO reorder ROIs in segmentation app by drag and drop
+- ROIs can be reordered in segmentation app by drag and drop
 - select ROI by clicking inside the polygon in the video playes (segmentation app)
 - Typos
 - Simplify idtrackerai/network file structure and imports